Update the qms protos to use doubles instead of floats
This results in quotas, quota defaults, and addons using float64 in the generated Go code instead of float64.

+++ java/org/cyverse/de/protobufs/AddonOrBuilder.java
@@ -99,10 +99,10 @@ public interface AddonOrBuilder extends
    * How much of the resource type is added to the quota by the add-on.
    * </pre>
    *
-   * <code>float default_amount = 5 [json_name = "default_amount"];</code>
+   * <code>double default_amount = 5 [json_name = "default_amount"];</code>
    * @return The defaultAmount.
    */
-  float getDefaultAmount();
+  double getDefaultAmount();
 
   /**
    * <pre>
